# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API Documentation for User API",
        "contact": {
            "email": "your-email@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Local Server"
        }
    ],
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get All Posts with published users",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "Posts retrieved successfully"
                    },
                    "404": {
                        "description": "No posts found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Store Post",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/{post_id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get a specific post",
                "operationId": "82662f799bed4e5592b7e1ce88970f5b",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post retrieved successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Update an existing post",
                "operationId": "d790ba4524d74e20234f3b6ec11ea198",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Delete a post",
                "operationId": "aa7c64fba1cae36381e525046298a7ad",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted successfully"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/deleted": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get all deleted posts",
                "operationId": "5e0089eb563e9786d7895dcfb0fa388c",
                "responses": {
                    "200": {
                        "description": "Deleted posts retrieved successfully"
                    },
                    "404": {
                        "description": "No deleted posts found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/restore/{post_id}": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Restore a deleted post",
                "operationId": "e97fa0e77c9eafeaaf0be52e0efa04f2",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post restored successfully"
                    },
                    "404": {
                        "description": "No deleted post found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/force_delete/{post_id}": {
            "delete": {
                "tags": [
                    "posts"
                ],
                "summary": "Permanently delete a post",
                "operationId": "3423c1daeeedbdffd2daec3e812550a3",
                "parameters": [
                    {
                        "name": "post_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post permanently deleted"
                    },
                    "404": {
                        "description": "Post not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/search": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Search for a post by title",
                "operationId": "42d93c0962544591249c10666536dd25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "search": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post Found"
                    },
                    "404": {
                        "description": "Post Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/posts/filter": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Filter and sort posts",
                "operationId": "3cecfb858b5f5af4c64332b08ecbf213",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "from_date": {
                                        "description": "The start date for filtering posts (format: d-m-Y)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "28-07-2024"
                                    },
                                    "to_date": {
                                        "description": "The end date for filtering posts (format: d-m-Y)",
                                        "type": "string",
                                        "format": "date",
                                        "example": "30-07-2024"
                                    },
                                    "is_active": {
                                        "description": "Filter posts based on their activity status",
                                        "type": "integer",
                                        "example": 0
                                    },
                                    "sort_by": {
                                        "description": "The field by which to sort posts",
                                        "type": "string",
                                        "enum": [
                                            "created_at",
                                            "is_active"
                                        ],
                                        "example": "created_at"
                                    },
                                    "sort_direction": {
                                        "description": "The direction in which to sort posts",
                                        "type": "string",
                                        "enum": [
                                            "asc",
                                            "desc"
                                        ],
                                        "example": "asc"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post Found"
                    },
                    "404": {
                        "description": "Post Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}